# lib/api-zod/src/generated/types/legalAcceptance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LegalAcceptanceKey } from "./legalAcceptanceKey";

export interface LegalAcceptance {
  key: LegalAcceptanceKey;
  /** @minimum 1 */
  versionId: number;
  accepted: boolean;
}