# lib/api-zod/src/generated/types/legalAcceptanceRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LegalAcceptanceRecordPolicyKey } from "./legalAcceptanceRecordPolicyKey";
import type { LegalAcceptanceRecordSource } from "./legalAcceptanceRecordSource";

export interface LegalAcceptanceRecord {
  userId: number;
  policyKey: LegalAcceptanceRecordPolicyKey;
  versionId: number;
  acceptedAt: Date;
  source: LegalAcceptanceRecordSource;
}