# lib/api-zod/src/generated/types/legalConsentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LegalAcceptance } from "./legalAcceptance";

export interface LegalConsentInput {
  /**
   * @minItems 1
   * @maxItems 3
   */
  acceptances: LegalAcceptance[];
}