# lib/api-zod/src/generated/types/legalConsentStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LegalPolicy } from "./legalPolicy";

export interface LegalConsentStatus {
  policies: LegalPolicy[];
  outstanding: LegalPolicy[];
  requiresAcceptance: boolean;
}