# lib/api-zod/src/generated/types/legalHistory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LegalAcceptanceRecord } from "./legalAcceptanceRecord";
import type { LegalPolicy } from "./legalPolicy";

export interface LegalHistory {
  versions: LegalPolicy[];
  acceptances: LegalAcceptanceRecord[];
  page: number;
  pageSize: number;
  total: number;
  versionTotal: number;
  acceptanceTotal: number;
}