# lib/api-zod/src/generated/types/legalPolicyManifest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LegalPolicy } from "./legalPolicy";

export interface LegalPolicyManifest {
  policies: LegalPolicy[];
}